Fixed 1 bajillion bugs. ALSO /INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK
/INTEGRITYCHECK

Also, CHECK if a list is empty, not if the first element is not NULL
IT POINTS TO ITSELF

Triggered

+++ Protect/Callbacks.c
@@ -21,7 +21,7 @@ UNICODE_STRING Altitude;
 BOOLEAN CallbackInstalled = FALSE;
 PVOID RegistrationHandle = NULL;
 
-UINT32 CurrentWatchCount = 0;
+ULONG CurrentWatchCount = 0;
 
 VOID
 RegisterCallbacks(
